added UID to PID mapping

+++ src/rozzy/proxy/shell.py
@@ -20,12 +20,14 @@ logger.setLevel(logging.DEBUG)
 class Popen:
     def __init__(self,
                  args: str,
+                 uid: str,
                  shell: 'ShellProxy',
                  api_docker: DockerAPIClient,
                  exec_id: int,
                  stream: Iterator[bytes]
                  ) -> None:
         self.__args = args
+        self.__uid = uid
         self.__shell = shell
         self.__api_docker = api_docker
         self.__exec_id = exec_id
@@ -45,12 +47,21 @@ class Popen:
     def args(self) -> str:
         return self.__args
 
+    def __uid_to_pid(self) -> Optional[int]:
+        prefix = f'/bin/bash -c echo {self.__uid} > /dev/null && '
+        cmd = f'ps -eo pid,cmd | grep "{prefix}"'
+        while not self.finished:
+            code, output, duration = self.__shell.execute(cmd)
+            for line in output.split('\n'):
+                pid_str, _, p_cmd = line.strip().partition(' ')
+                if p_cmd.startswith(prefix):
+                    return int(pid_str)
+        return None
+
     @property
     def pid(self) -> Optional[int]:
-        if not self.__pid:
-            pid = self._inspect()['Pid']
-            if pid != 0:
-                self.__pid = pid
+        if not self.__pid and not self.finished:
+            self.__pid = self.__uid_to_pid()
         return self.__pid
 
     @property
@@ -64,12 +75,9 @@ class Popen:
         return self.__returncode
 
     def send_signal(self, sig: int) -> None:
-        # FIXME this is the host PID!
         pid = self.pid
-        while not pid:
-            pid = self.pid
-        logger.debug("killing process %d", pid)
-        self.__shell.send_signal(pid, sig)
+        if pid:
+            self.__shell.send_signal(pid, sig)
 
     def kill(self) -> None:
         self.send_signal(signal.SIGKILL)
@@ -92,14 +100,20 @@ class ShellProxy:
     def send_signal(self, pid: int, sig: int) -> None:
         self.execute(f'kill -{sig} {pid}', user='root')
 
+    def __generate_popen_uid(self) -> str:
+        return "cool"
+
     def instrument(self,
                    command: str,
                    time_limit: Optional[int] = None,
-                   kill_after: int = 1
+                   kill_after: int = 1,
+                   identifier: Optional[str] = None
                    ) -> str:
         logger.debug("instrumenting command: %s", command)
         q = shlex.quote
         command = f'source /.environment && {command}'
+        if identifier:
+            command = f'echo {q(identifier)} > /dev/null && {command}'
         command = f'/bin/bash -c {q(command)}'
         if time_limit:
             command = (f'timeout --kill-after={kill_after} '
@@ -117,17 +131,19 @@ class ShellProxy:
               time_limit: Optional[int] = None,
               kill_after: int = 1
               ) -> Popen:
+        uid_popen = self.__generate_popen_uid()
         id_container = self.__container_docker.id
         api_docker = self.__api_docker
         command_orig = command
-        command = self.instrument(command, time_limit, kill_after)
+        command = self.instrument(command, time_limit, kill_after,
+                                  identifier=uid_popen)
         exec_resp = api_docker.exec_create(id_container, command,
                                            tty=True,
                                            stdout=stdout,
                                            stderr=stderr)
         exec_id = exec_resp['Id']
         stream = api_docker.exec_start(exec_id, stream=True, tty=True)
-        return Popen(command_orig, self, api_docker, exec_id, stream)
+        return Popen(command_orig, uid_popen, self, api_docker, exec_id, stream)
 
     def execute(self,
                 command: str,
